Making project have a real dep setup
Since nancy switched over to using dep to actually do the parsing this
project no longer worked when being used. So create a more realistic
project.

We dont actually use the dependencies but it still creates something
that is vunerable since the versions are pinned to deps that have
vunerabilities