Build script stuff....
Making it so we run through more paths for now in here. All of them
should exit with non 0 exit codes.

Making the build pass as long as exit code != 0 for all of them.

If any one suceededs that means this project is no longer vulnerable or
nancy is broken and that is not what we want.